Changes to registration page
