+++ client/src/HEDemonstration/HEDemo.css
@@ -9,7 +9,7 @@
     margin: 0 auto;
 }
   
-h1 {
+.h1 {
     font-size: 24px;
     margin-bottom: 10px;
     margin-top: 80px;
